Embed dialogue act labels(distribution) in model

+++ onmt/inputters/inputter.py
@@ -134,8 +134,10 @@ def get_fields(src_data_type, n_src_features, n_tgt_features, n_knl_features):
     # using Dialogue Act Label
     # TODO: apply model_mode (accept as argument)
     if src_data_type == 'text':
-        fields["src_da_label"] = Field(sequential=False, use_vocab=False)
-        fields["tgt_da_label"] = Field(sequential=False, use_vocab=False)
+        #fields["src_da_label"] = Field(sequential=False, use_vocab=False)
+        #fields["tgt_da_label"] = Field(sequential=False, use_vocab=False)
+        fields["src_da_label"] = Field(sequential=False, use_vocab=False, dtype=torch.float)
+        fields["tgt_da_label"] = Field(sequential=False, use_vocab=False, dtype=torch.float)
 
     return fields
